okay, a couple more changes
Slight tweaks/edits to graphics text after looking at the app on the
iPhone 4s (target display size).  Know what’s cool?  The new slider on
Activity Monitor still works on the 4s, way at the bottom of the
screen, barely visible.  It adjusts sensitivity of both the graphics
display and the shuffle/swipe counting, relative to actual motion of
the Node.