Reworked the Side-Side/Front-Back counting
Counters now increment in a predictable way that matches the meters…
that is, when you overdrive one or the other meter by a little bit,
that counter increments.  It uses a simplified peak-detection scheme to
decide when to register a movement, and also to decide which one to
increment.  The user can see (and learn to predict) how much movement
will increment a counter, by looking at how much acceleration saturates
the meter, and adjust the slider for comfort (easy) level or more
challenge.  Works best when combined with updated firmware from
Variable (currently in beta only) which eliminates response lag in the
accel. data.